requirements_for_Qwen.txt
accelerate==1.12.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
astor==0.8.1
async-timeout==5.0.1
attrdict==2.0.1
attrs==26.1.0
av==16.0.1
babel==2.17.0
basicsr==1.4.2
bce-python-sdk==0.9.59
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
conda-pack==0.7.1
contourpy==1.3.2
controlnet_aux==0.0.7
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
decord==0.6.0
einops==0.7.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fairscale==0.4.13
fastapi==0.124.2
ffmpy==1.0.0
filelock==3.19.1
fire==0.7.1
flash-attn==2.7.2.post1
Flask==3.1.2
flask-babel==4.0.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gradio==5.4.0
gradio_client==1.4.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
imagesize==1.4.1
imgaug==0.4.0
importlib_metadata==9.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
lightning==2.6.1
lightning-utilities==0.15.3
lmdb==1.7.5
loralib==0.1.2
lpips==0.1.4
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.33.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-ml-py==13.595.45
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
opencv-contrib-python==4.6.0.66
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.5
packaging==25.0
paddleocr==2.7.0.3
paddlepaddle==2.6.2
paddlex==3.3.13
pandas==2.3.3
pdf2docx==0.5.8
peft==0.17.0
pillow==11.3.0
pillow_heif==1.3.0
platformdirs==4.9.4
polars==1.37.1
polars-runtime-32==1.37.1
premailer==3.10.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.4
psutil==7.1.3
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.20.2
pynvml==13.0.1
pyparsing==3.3.1
pypdfium2==5.3.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.12
pytorch-lightning==2.6.1
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.10
RapidFuzz==3.14.3
rarfile==4.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.19.1
ruff==0.14.9
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.1
starlette==0.50.0
sympy==1.14.0
termcolor==3.3.0
tifffile==2025.5.10
timm==1.0.24
tokenizers==0.22.2
tomli==2.4.1
tomlkit==0.12.0
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchmetrics==1.9.0
torchvision==0.19.0+cu124
tqdm==4.67.1
transformers==4.56.0
transformers-stream-generator==0.0.4
triton==3.0.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.4.5
ultralytics-thop==2.0.18
urllib3==2.6.2
uvicorn==0.38.0
visualdl==2.5.3
wcwidth==0.2.14
websockets==12.0
Werkzeug==3.1.5
yapf==0.43.0
yarl==1.23.0
zipp==3.23.0
